Baubas here
Baubas, an evil spirit with long lean arms, wrinkly fingers and red eyes. He harasses people and tears their hair or stifles them. To children, he is the equivalent of the boogeyman of the English-speaking countries. A misbehaving child could be told by the parents: "Behave, or baubas will come and get you". Also it could be described as a black and dark creature living under the carpet or in some dark spot of the house.